Crate mux

Data structures representing the mux protocol.

Package mux implements a generic RPC multiplexer with a rich protocol. Mux is itself encoding independent, so it is meant to use as the transport for other RPC systems (eg. thrift). In OSI terminology, it is a pure session layer.

Modules

codec
Codec functions for encoding and decoding mux frames.

types
Wire identification tag of the mux message types.

Structs

Dentry
Single entry of the Dtab

Dtab
Delegate table.

Init
Representation of a mux Tinit and Rinit frame.

Message
Representation of an entire mux packet.

Rdispatch
Representation of a mux Rdispatch frame.

Rerr
Representation of a mux Rerr frame.

Tag
Id number and end flag for message frames.

Tdiscarded
Representation of a mux Tdiscarded frame.

Tdispatch
Representation of a mux Tdispatch frame.

Tlease
Representation of a mux Tlease frame.

Treq
Representation of the mux Treq types.

Enums

MessageFrame
Type wrapper for the mux packet representations.

Rmsg
Representation of a mux Rreq and Rdispatch message body.

Constants

MAX_TAG
Maximum value of a mux Tag

Type Definitions

Contexts
Contexts of dispatch and init messages.

Headers
Headers for a Treq.