Module mux::codec

Codec functions for encoding and decoding mux frames.

Decode messages from a continuous stream using read_message to address message framing. Individual message decoding functions assume the remainder of the stream comprises the message.

Modules

size

Functions

decode_contexts

decode_dtab

decode_frame
Synchronously decode a mux MessageFrame

decode_headers

decode_init

decode_message
Synchronously decode a mux Message

decode_rdispatch

decode_rerr

decode_rreq

decode_tag

decode_tdiscarded

decode_tdispatch

decode_tlease

decode_treq

decode_u16_string

encode_contexts

encode_dtab

encode_frame
Synchronously encode a MessageFrame to the Write

encode_headers

encode_init

encode_message
Synchronously encode a Message to the Write

encode_rdispatch

encode_rerr

encode_rreq

encode_tag

encode_tdiscarded

encode_tdispatch

encode_tlease

encode_treq

encode_u16_string

read_message
Synchronously read a whole mux Message

write_message
Synchronously encode a Message to the Write with the frame size