Function mux::codec::decode_frame

pub fn decode_frame<R: Read>(tpe: i8, reader: R) -> Result<MessageFrame>

Synchronously decode a mux MessageFrame

This function will synchronously read from the provided &mut Read until it has decoded a frame. Frame length is assumed to be triggered by an EOF. When using use a continuous stream, such as a TcpStream, use Read.take(n) if you know the frame length or use read_message above to decode an entire Message and then extract the frame.

use std::io::Cursor;
use mux::MessageFrame;
use mux::codec;
use mux::types;

let mut r = Cursor::new(vec![]); // ping frame is empty
let frame = codec::decode_frame(types::TPING, &mut r).unwrap();
assert_eq!(frame, MessageFrame::Tping);